+++ package.json
@@ -30,5 +30,10 @@
     "lowdb": "^1.0.0",
     "node-html-parser": "^6.1.5",
     "winston": "^3.10.0"
+  },
+  "devDependencies": {
+    "@types/lowdb": "^1.0.11",
+    "@types/node": "^20.5.0",
+    "typescript": "^5.1.6"
   }
 }
