+++ src/models/UserData.ts
@@ -1,0 +1,5 @@
+export interface UserData {
+    user_id: String;
+    csrf_token: String;
+
+}
