+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "outDir": "./build",
+    "allowJs": true,
+    "target": "ESNext",
+    "sourceMap": true,
+    "module": "CommonJS",
+    "allowSyntheticDefaultImports": true
+  },
+  "include": ["./src/**/*"]
+}
